Initial Pause Button Push
Made start of pause button but was not able to finish. Pause button can be pulled up or exited with escape. It pauses the scene as is. You can quit the pause button with escape or by clicking resume game. I will add a quit and restart button after class.